Add ember-a11y-testing with ability to toggle a11yAudit acceptance tests
