+++ package.json
@@ -26,6 +26,7 @@
     "algoliasearch": "^3.30.0",
     "broccoli-asset-rev": "^3.0.0",
     "broccoli-static-site-json": "^2.1.2",
+    "ember-a11y-testing": "^0.5.7",
     "ember-body-class": "^1.1.0",
     "ember-cli": "~3.8.1",
     "ember-cli-app-version": "^3.0.0",
